backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
markdown==3.5.2
bleach==6.1.0
aiofiles==23.2.1
httpx==0.26.0
pypinyin==0.51.0
rarfile==4.2
py7zr==0.21.0
supabase==2.5.0
boto3==1.34.0
